Move VerifiesAssignmentStudent to CreatesGrade
- VerifiesAssignmentStudent is no longer specifically for criterion grades so it should be moved to creates_grade directory
- Explicit require statements added to creates_grade to stay consistent
