+++ app/services/creates_grade.rb
@@ -1,4 +1,10 @@
 require "light-service"
+require_relative "creates_grade/verifies_assignment_student"
+require_relative "creates_grade/builds_grade"
+require_relative "creates_grade/associates_submission_with_grade"
+require_relative "creates_grade/marks_as_graded"
+require_relative "creates_grade/saves_grade"
+require_relative "creates_grade/runs_grade_updater_job"
 
 module Services
   class CreatesGrade
